fix pose cov to pose
