+++ aichallenge/workspace/src/aichallenge_submit/pose_transformer/package.xml
@@ -1,6 +1,6 @@
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>pose_cov_transformer</name>
+  <name>pose_transformer</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="yoshi-22@todo.todo">yoshi-22</maintainer>
